+++ lib/handler/authorizationHandler.js
