+++ lib/index.js
@@ -51,7 +51,7 @@ var log = logMaker('lib/index.js');
     config,
     messageApi,
     userApiClient,
-    require('./handler/gatekeeperHandler')(crudHandler)
+    require('./handler/authorizationHandler')(crudHandler)
   );
 
   lifecycle.add('server', messagesService);
