+++ lib/messagesService.js
@@ -22,7 +22,7 @@ var log = require('./log.js')('messagesService.js');
 
 var server, servicePort;
 
-var messagesService = function(envConfig, messageApi, userApiClient, gatekeeperHandler) {
+var messagesService = function(envConfig, messageApi, userApiClient, authorizationHandler) {
   //create the server depending on the type
   if (envConfig.httpPort != null) {
     servicePort = envConfig.httpPort;
@@ -30,7 +30,7 @@ var messagesService = function(envConfig, messageApi, userApiClient, gatekeeperH
       { name: 'TidepoolMessageHttp' },
       messageApi,
       userApiClient,
-      gatekeeperHandler
+      authorizationHandler
     );
   }
 
@@ -40,7 +40,7 @@ var messagesService = function(envConfig, messageApi, userApiClient, gatekeeperH
       _.extend({ name: 'TidepoolMessageHttps'}, envConfig.httpsConfig),
       messageApi,
       userApiClient,
-      gatekeeperHandler
+      authorizationHandler
     );
   }
 
@@ -51,7 +51,7 @@ var messagesService = function(envConfig, messageApi, userApiClient, gatekeeperH
 };
 
 
-function createServer(serverConfig, messageApi, userApiClient, gatekeeperHandler){
+function createServer(serverConfig, messageApi, userApiClient, authorizationHandler){
 
   log.info('Creating server[%s]', serverConfig.name);
   server = restify.createServer(serverConfig);
@@ -68,21 +68,21 @@ function createServer(serverConfig, messageApi, userApiClient, gatekeeperHandler
   server.get('/status', messageApi.status);
 
   //finding of message(s)
-  server.get('/read/:msgid', checkToken, gatekeeperHandler.checkPermissionsFromMessage, messageApi.findById);
-  server.get('/thread/:msgid', checkToken, gatekeeperHandler.checkPermissionsFromMessage, messageApi.getThread);
+  server.get('/read/:msgid', checkToken, authorizationHandler.authorizeFromMessage, messageApi.findById);
+  server.get('/thread/:msgid', checkToken, authorizationHandler.authorizeFromMessage, messageApi.getThread);
 
-  server.get('/all/:groupid', checkToken, gatekeeperHandler.checkPermissionsFromGroup, messageApi.findAllById);
-  server.get('/notes/:groupid', checkToken, gatekeeperHandler.checkPermissionsFromGroup, messageApi.getNotes);
+  server.get('/all/:groupid', checkToken, authorizationHandler.authorize, messageApi.findAllById);
+  server.get('/notes/:groupid', checkToken, authorizationHandler.authorize, messageApi.getNotes);
 
   //adding messages
-  server.post('/send/:groupid', checkToken, gatekeeperHandler.checkPermissionsFromGroup, messageApi.addThread);
-  server.post('/reply/:msgid', checkToken, gatekeeperHandler.checkPermissionsFromMessage, messageApi.replyToThread);
+  server.post('/send/:groupid', checkToken, authorizationHandler.authorize, messageApi.addThread);
+  server.post('/reply/:msgid', checkToken, authorizationHandler.authorizeFromMessage, messageApi.replyToThread);
 
   //edit
-  server.put('/edit/:msgid', checkToken, gatekeeperHandler.checkPermissionsFromMessage, messageApi.editMessage);
+  server.put('/edit/:msgid', checkToken, authorizationHandler.authorizeFromMessage, messageApi.editMessage);
 
   //delete
-  server.del('/remove/:msgid', checkToken, gatekeeperHandler.checkPermissionsFromMessage, messageApi.removeMessage);
+  server.del('/remove/:msgid', checkToken, authorizationHandler.authorizeFromMessage, messageApi.removeMessage);
 }
 
 function stopService() {
